% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gdfpd_utils.R
\name{gdfpd.convert.to.wide}
\alias{gdfpd.convert.to.wide}
\title{Converts a dataframe from gdpfd_GetDFPData to the wide format}
\usage{
gdfpd.convert.to.wide(data.in)
}
\arguments{
\item{data.in}{Data frame with financial information}
}
\value{
A dataframe in the wide format
}
\description{
Converts a dataframe from gdpfd_GetDFPData to the wide format
}
\examples{

\dontrun{
name.companies <- 'PETROBRAS'
first.date <- '2000-01-01'
last.date <-  '2005-01-01'
df.statements <- gdfpd.GetDFPData(name.companies = name.companies,
                                 first.date = first.date,
                                 last.date = last.date)

df.assets <- df.statements$assets[[1]]
df.assets.wide <- gdfpd.convert.to.wide(df.assets)
}
}
