% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gdfpd_export_DFP_Data.R
\name{gdfpd.export.DFP.data}
\alias{gdfpd.export.DFP.data}
\title{Export tibble to an excel file}
\usage{
gdfpd.export.DFP.data(data.in, base.file.name = paste0("GetDFPData_Export_",
  Sys.Date()), type.export = "xlsx", format.data = "wide")
}
\arguments{
\item{data.in}{Tibble with financial information (output of gdfpd.GetDFPData)}

\item{base.file.name}{The basename of excel file (don't include extension)}

\item{type.export}{The extension of the desired format (only 'xlsx' implemented so far)}

\item{format.data}{The format of output. 'wide' (default) for a wide table and 'long' for a long table}
}
\value{
nothing
}
\description{
Export tibble to an excel file
}
\examples{
\dontrun{
name.companies <- 'PETROBRAS'
first.date <- '2005-01-01'
last.date <-  '2006-01-01'

df.statements <- gdfpd.GetDFPData(name.companies = name.companies,
                                  first.date = first.date,
                                  last.date = last.date)

gdfpd.export.DFP.data(df.statements, format.data = 'wide')
}
}
