% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gdfpd_get_info_companies.R
\name{gdfpd.get.info.companies}
\alias{gdfpd.get.info.companies}
\title{Reads up to date information about Bovespa companies from a github file}
\usage{
gdfpd.get.info.companies()
}
\value{
A dataframe with several information about Bovespa companies
}
\description{
A csv file with information about available companies and time periods is downloaded from github and read.
This file is updated periodically and manually by the author.
}
\examples{

\dontrun{ # make it fast to run in CRAN
df.info <- gdfpd.get.info.companies()
str(df.info)
}
}
