% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gdfpd_GetDFPData.R
\name{gdfpd.GetDFPData}
\alias{gdfpd.GetDFPData}
\title{Downloads and reads Title}
\usage{
gdfpd.GetDFPData(name.companies, first.date, last.date,
  folder.out = tempdir())
}
\arguments{
\item{folder.out}{}
}
\description{
Downloads and reads Title
}
