% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gdfpd_read_zip_file.R
\name{gdfpd.read.zip.file}
\alias{gdfpd.read.zip.file}
\title{Reads a single zip file with financial data for a trimester}
\usage{
gdfpd.read.zip.file(my.zip.file, folder.out = tempdir())
}
\arguments{
\item{my.zip.file}{Full path to zip file}

\item{folder.out}{Folder to unzip files (default = tempdir())}
}
\value{
A list with several dataframes containing financial statements
}
\description{
Reads a single zip file with financial data for a trimester
}
\examples{
# no example
}
