% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gdfpd_read_zip_file.R
\name{gdfpd.read.zip.file.type.1}
\alias{gdfpd.read.zip.file.type.1}
\title{Reads zip file post 2011 (internal)}
\usage{
gdfpd.read.zip.file.type.1(my.zip.file, folder.to.unzip = tempdir())
}
\arguments{
\item{my.zip.file}{Full path to zip file}

\item{folder.to.unzip}{Folder to unzip files (default = tempdir())}
}
\value{
A list with financial statements
}
\description{
Reads zip file post 2011 (internal)
}
\examples{
# no example
}
