% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gdfpd_utils.R
\name{gdfpd.search.company}
\alias{gdfpd.search.company}
\title{Helps users search for a company name}
\usage{
gdfpd.search.company(char.to.search)
}
\arguments{
\item{char.to.search}{Character for partial matching}
}
\value{
Names of found companies
}
\description{
Helps users search for a company name
}
\examples{

\dontrun{
df.info <- gdfpd.search.company('GERDAU')
}
}
