% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gitrd_Bovespa_Data.R
\name{gitrd.get.bovespa.data}
\alias{gitrd.get.bovespa.data}
\title{Reads information for a company from B3 site}
\usage{
gitrd.get.bovespa.data(my.id)
}
\arguments{
\item{my.id}{A CVM id}
}
\value{
A list with several dataframes
}
\description{
Given a CVM code, this function scrapes information from the company page.
}
\examples{

\dontrun{ # keep cran check fast
l.info.PETR <- gitrd.get.dovespa.data(my.id = 9512)
str(l.info.PETR)
}
}
