% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gitrd_get_inflation_data.R
\name{gitrd.get.inflation.data}
\alias{gitrd.get.inflation.data}
\title{Downloads and read inflation data from github}
\usage{
gitrd.get.inflation.data(inflation.index)
}
\arguments{
\item{inflation.index}{Sets the inflation index to use for finding inflation adjusted values of all reports. Possible values: 'dollar' (default) or 'IPCA', the brazilian main inflation index.
When using 'IPCA', the base date is set as the last date found in itr/dfp dataset.}
}
\value{
A dataframe with inflation data
}
\description{
Inflation data is available at git repo 'msperlin/GetITRData_auxiliary'
}
\examples{

\dontrun{ # keep cran check fast
df.inflation <- gitrd.get.inflation.data('IPCA')
str(df.inflation)
}
}
