% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gitrd_utils.R
\name{gitrd.read.fwf.file}
\alias{gitrd.read.fwf.file}
\title{Reads FWF file from bovespa (internal)}
\usage{
gitrd.read.fwf.file(my.f, type.fin.report)
}
\arguments{
\item{my.f}{File to be read}

\item{type.fin.report}{Peridodicy of fin data ('itr' - quarterly or 'dfp' - annual)}
}
\value{
A dataframe with data
}
\description{
Reads FWF file from bovespa (internal)
}
\examples{

my.f <- system.file('extdata/ITRBPAE.001', package = 'GetITRData')

df.assets <- gitrd.read.fwf.file(my.f, type.fin.report = 'itr')
}
