% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gitrd_utils.R
\name{gitrd.search.company}
\alias{gitrd.search.company}
\title{Helps users search for a company name}
\usage{
gitrd.search.company(char.to.search)
}
\arguments{
\item{char.to.search}{Character for partial matching}
}
\value{
Names of found companies
}
\description{
Helps users search for a company name
}
\examples{

\dontrun{
df.info <- gitrd.search.company('GERDAU')
}
}
