% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gdfpd_get_info_companies.R
\name{gdfpd.get.info.companies}
\alias{gdfpd.get.info.companies}
\title{Reads up to date information about Bovespa companies from a github file}
\usage{
gdfpd.get.info.companies()
}
\value{
A dataframe with information about Bovespa companies
}
\description{
Reads up to date information about Bovespa companies from a github file
}
\examples{

df.info <- gdfpd.get.info.companies()
}
