% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gitrd_utils.R
\name{gitrd.convert.to.wide}
\alias{gitrd.convert.to.wide}
\title{Converts a dataframe from gdpfd_GetITRData to the wide format}
\usage{
gitrd.convert.to.wide(data.in)
}
\arguments{
\item{data.in}{Data frame with financial information}
}
\value{
A dataframe in the wide format
}
\description{
Converts a dataframe from gdpfd_GetITRData to the wide format
}
\examples{

\dontrun{
name.companies <- 'PETROBRAS'
first.date <- '2000-01-01'
last.date <-  '2005-01-01'
df.statements <- gitrd.GetITRData(name.companies = name.companies,
                                 first.date = first.date,
                                 last.date = last.date)

df.assets <- df.statements$assets[[1]]
df.assets.wide <- gitrd.convert.to.wide(df.assets)
}
}
