% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gitrd_utils.R
\name{gitrd.fix.dataframes}
\alias{gitrd.fix.dataframes}
\title{Fix dataframe for version issues (internal)}
\usage{
gitrd.fix.dataframes(df.in)
}
\arguments{
\item{df.in}{A dataframe with financial statements}
}
\value{
The fixed data.frame
}
\description{
Fix dataframe for version issues (internal)
}
\examples{
 # no example
}
