% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gitrd_utils.R
\name{gitrd.read.fwf.file}
\alias{gitrd.read.fwf.file}
\title{Reads FWF file from bovespa (internal)}
\usage{
gitrd.read.fwf.file(my.f)
}
\arguments{
\item{my.f}{File to be read}
}
\value{
A dataframe with data
}
\description{
Reads FWF file from bovespa (internal)
}
\examples{
# no example
}
