% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gitrd_read_zip_file.R
\name{gitrd.read.zip.file.type.1}
\alias{gitrd.read.zip.file.type.1}
\title{Reads folder for zip file post 2011 (internal)}
\usage{
gitrd.read.zip.file.type.1(rnd.folder.name, folder.to.unzip = tempdir())
}
\arguments{
\item{rnd.folder.name}{Folder where unzipped files are available}

\item{folder.to.unzip}{Folder to unzip files (default = tempdir())}
}
\value{
A list with financial statements
}
\description{
Reads folder for zip file post 2011 (internal)
}
\examples{
# no example (this functions not used directly)
}
